Subjectを機能ごとに分割
